+++ packages/fluentui-rc/src/CheckboxWidget/CheckboxWidget.tsx
@@ -54,7 +54,7 @@ export default function CheckboxWidget<
 
   return (
     <>
-      {!hideLabel && !!description && (
+      {!hideLabel && description && (
         <DescriptionFieldTemplate
           id={descriptionId<T>(id)}
           description={description}