+++ packages/mui/src/DescriptionField/DescriptionField.tsx
@@ -1,5 +1,6 @@
 import Typography from '@mui/material/Typography';
 import { DescriptionFieldProps, FormContextType, RJSFSchema, StrictRJSFSchema } from '@rjsf/utils';
+import { RichDescription } from '@rjsf/core';
 
 /** The `DescriptionField` is the template to use to render the description of a field
  *
@@ -10,11 +11,11 @@ export default function DescriptionField<
   S extends StrictRJSFSchema = RJSFSchema,
   F extends FormContextType = any,
 >(props: DescriptionFieldProps<T, S, F>) {
-  const { id, description } = props;
+  const { id, description, registry, uiSchema } = props;
   if (description) {
     return (
       <Typography id={id} variant='subtitle2' style={{ marginTop: '5px' }}>
-        {description}
+        <RichDescription description={description} registry={registry} uiSchema={uiSchema} />
       </Typography>
     );
   }