+++ packages/react-bootstrap/src/DescriptionField/DescriptionField.tsx
@@ -1,19 +1,24 @@
 import { DescriptionFieldProps, FormContextType, RJSFSchema, StrictRJSFSchema } from '@rjsf/utils';
+import { RichDescription } from '@rjsf/core';
 
+/** The `DescriptionField` is the template to use to render the description of a field
+ *
+ * @param props - The `DescriptionFieldProps` for this component
+ */
 export default function DescriptionField<
   T = any,
   S extends StrictRJSFSchema = RJSFSchema,
   F extends FormContextType = any,
->({ id, description }: DescriptionFieldProps<T, S, F>) {
-  if (description) {
-    return (
-      <div>
-        <div id={id} className='mb-3'>
-          {description}
-        </div>
-      </div>
-    );
+>({ id, description, registry, uiSchema }: DescriptionFieldProps<T, S, F>) {
+  if (!description) {
+    return null;
   }
 
-  return null;
+  return (
+    <div>
+      <div id={id} className='mb-3'>
+        <RichDescription description={description} registry={registry} uiSchema={uiSchema} />
+      </div>
+    </div>
+  );
 }