+++ packages/snapshot-tests/src/objectTests.tsx
@@ -4,6 +4,12 @@ import { FormProps } from '@rjsf/core';
 import { RJSFSchema, UiSchema } from '@rjsf/utils';
 import validator from '@rjsf/validator-ajv8';
 
+jest.mock('@rjsf/utils', () => ({
+  ...jest.requireActual('@rjsf/utils'),
+  // Disable the getTestIds within the snapshot tests by returning an empty object
+  getTestIds: jest.fn(() => ({})),
+}));
+
 const titleAndDesc = {
   title: 'Test field',
   description: 'a test description',