+++ CHANGELOG.md
@@ -21,6 +21,58 @@ should change the heading of the (upcoming) version to include a major version b
 
 - Fixed issue with schema references in combinators(allOf, anyOf, oneOf) could not be modified when defaults were set, fixing [#4555](https://github.com/rjsf-team/react-jsonschema-form/issues/4555)
 
+# 6.0.0-beta.2
+
+## @rjsf/antd
+
+- Updated `DescriptionField` to render the `description` using the `RichDescription` field
+
+## @rjsf/chakra-ui
+
+- Updated `DescriptionField` to render the `description` using the `RichDescription` field
+
+## @rjsf/core
+
+- Added new `RichDescription` component, refactored from `SchemaField` to support Rich Text descriptions in Markdown format
+- Updated `DescriptionField` to render the `description` using the `RichDescription` field
+
+## @rjsf/daisyui
+
+- Updated `DescriptionField` to render the `description` using the `RichDescription` field
+- Updated `FieldTemplate` to move the checkbox implementation into the `CheckboxWidget` adding the `description` for checkboxes
+- Updated `package.json` to make the package publishable
+- Updated `DaisyUIFrameProvider` to extract the bulk of the code into `DaisyUIFrameComponent` to add a `useEffect()` with a cleanup to remove the tailwind styles
+
+## @rjsf/fluentui-rc
+
+- Updated `DescriptionField` to render the `description` using the `RichDescription` field
+
+## @rjsf/mui
+
+- Updated `DescriptionField` to render the `description` using the `RichDescription` field
+
+## @rjsf/react-bootstrap
+
+- Updated `DescriptionField` to render the `description` using the `RichDescription` field
+- Updated `CheckboxField` to remove the `checkbox` class that breaks the UI
+
+## @rjsf/semantic-ui
+
+- Updated `DescriptionField` to render the `description` using the `RichDescription` field
+
+## @rjsf/shadcn
+
+- Updated `DescriptionField` to render the `description` using the `RichDescription` field
+
+## @rjsf/utils
+
+- Updated the `description` field in field props to be a `string | ReactElement` and added `enableMarkdownInDescription` to the `GlobalUISchemaOptions` interface
+
+## Dev / docs / playground
+
+- Updated the `snapshot-tests` to disable `getTestId()` for snapshots and updated the `formTests.tsx` to add tests for rich text descriptions for generic fields and the `CheckboxWidget`
+- Updated the `uiSchema.md` to document new `enableMarkdownInDescription` prop
+- Updated the `playground` to move `daisyui` theme choice after `chakra-ui` and to stop freezing the samples to avoid an `AJV` validation issue
 
 # 6.0.0-beta.1